refactor(payments): slice payment services
